flatten schema module
