--- quiz/schema/__init__.py
@@ -1,6 +1,0 @@
-from .raw import load  # noqa
-from .to_types import *  # noqa
-
-__all__ = [
-    'Schema'
-]
